Remove started channel
Started channel was meant to indicate that the supervisor has started
all services. It was never accurate, and more importantly it was
wrong. The right way to know whether a service has been started is
to check the service itself.

In many tests, Simpleservice and friends have been replaced with
holdingservice, which is a service to whom you can ask whether it
has been started. In practice Simpleservice is used only for
examples.